Fetch messagesSentPerMinute for SQS and update it

+++ pkg/apis/workerpodautoscaler/v1alpha1/types.go
@@ -23,7 +23,7 @@ type WorkerPodAutoScalerSpec struct {
 	QueueURI                string   `json:"queueURI"`
 	DeploymentName          string   `json:"deploymentName"`
 	TargetMessagesPerWorker *int32   `json:"targetMessagesPerWorker"`
-	SecondsToProcessOneJob  *float32 `json:"secondsToProcessOneJob"`
+	SecondsToProcessOneJob  *float64 `json:"secondsToProcessOneJob"`
 	Disruptable             bool     `json:"disruptable"`
 }
